SVN: merge with branch 2.7/2.2 (rev.6646-6821)
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@6822 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/util/DataGenerator.java
@@ -77,7 +77,7 @@ public class DataGenerator extends Random {
 	 * @param max
 	 * @return random integer
 	 */
-	public long nextInt(int min, int max) {
+	public int nextInt(int min, int max) {
 		// raw random number
 		int r = Math.abs(nextInt());
